Fixed Compiler Errors

+++ Assets/Scripts/Networked Player Creation/NetState.cs
@@ -4,18 +4,18 @@ using System.Collections;
 public class NetState : MonoBehaviour
 {
 
-	public float timeStamp;
+	public double timeStamp;
 	public Vector3 pos;
 	public Quaternion rot;
 	
 	public NetState()
 	{
-		timeStamp = 0f;
+		timeStamp = 0d;
 		pos = Vector3.zero;
 		rot = Quaternion.identity;
 	}
 	
-	public NetState(float time, Vector3 pos, Quaternion rot)
+	public NetState(double time, Vector3 pos, Quaternion rot)
 	{
 		timeStamp = time;
 		this.pos = pos;
